requirements.txt
flask
requests
beautifulsoup4
googlesearch-python
fake-useragent
urllib3
numpy
rank-bm25
gunicorn
pandas
validators
openai